Выносит компонент ChallengeItem в отдельный файл
